+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # ActionLoggerBundle Changelog
 
 ## [Unreleased 0.2.0]
+### Add
+- New `EntityDeleteAction`
+
 ### Change
 - **BC-BREAK** Add keyHash to LogActionRelation used to find related log actions.
 - **BC-BREAK** Rename table log to log_action and log_relations to log_action_relation